Updates v1 network descriptor to act as a set of Organizations
