+++ fabctl-sandbox/src/main/java/org/hyperledger/fabric/fabctl/v1/network/OrdererConfig.java
@@ -12,4 +12,10 @@ public class OrdererConfig
 {
     public final String name;
     public final Environment environment;
+
+    public OrdererConfig(final String name, final Environment environment)
+    {
+        this.name = name;
+        this.environment = environment;
+    }
 }
