+++ fabctl-sandbox/src/main/java/org/hyperledger/fabric/fabctl/v1/network/OrganizationConfig.java
@@ -1,0 +1,26 @@
+/*-
+ * Copyright IBM Corp. All Rights Reserved.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package org.hyperledger.fabric.fabctl.v1.network;
+
+import java.lang.reflect.Array;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.Data;
+
+@Data
+public class OrganizationConfig
+{
+    public final String name;
+    public String mspID;
+
+    public final List<PeerConfig> peers = new ArrayList<>();
+    public final List<OrdererConfig> orderers = new ArrayList<>();
+
+    public OrganizationConfig(final String name)
+    {
+        this.name = name;
+    }
+}
